# quiniela/app/api/apidoc/models.go
// Package apidoc holds types referenced by swag annotations for OpenAPI schemas.
package apidoc

// HealthResponse is returned by GET /v1/healthz.
type HealthResponse struct {
	Status string `json:"status"`
}

// ErrorResponse is a generic JSON error body.
type ErrorResponse struct {
	Error string `json:"error"`
}

// CreateUserRequest is the body for POST /v1/users.
type CreateUserRequest struct {
	Name string `json:"name"`
}

// CreateUserResponse is returned on successful user creation.
type CreateUserResponse struct {
	Message string `json:"message"`
	Name    string `json:"name"`
}

// PostPredictionRequest is the body for POST /v1/predictions.
type PostPredictionRequest struct {
	FixtureID int32 `json:"fixture_id"`
	UserID    int32 `json:"user_id"`
	PredA     int32 `json:"pred_a"`
	PredB     int32 `json:"pred_b"`
}

// UserScoreResponse is returned by GET /v1/users/{userID}/score.
type UserScoreResponse struct {
	TotalPoints int64 `json:"total_points"`
}